separarating veiw from container components

+++ app/components/scenes/myJobs/Future.js
@@ -1,55 +1,8 @@
 import React from "react";
-import { Col, Table } from "reactstrap";
 import PropTypes from "prop-types";
-import moment from "moment";
-import firebase from "firebase";
-import { firestore, auth } from "../../../firebase";
-import { collectIdsAndDocs, formatDate, formatTime } from "../../../utilities";
-
-const FullTable = props => {
-  const { jobs, handleRemove } = props;
-  return (
-    <div className="container-fluid">
-      <div className="row">
-        <div className="col content shadow">
-          <h5 className="tableHeading">Upcomping Jobs</h5>
-          <Table hover striped responsive>
-            <thead>
-              <tr>
-                <th>Date</th>
-                <th>Time</th>
-                <th>Client</th>
-                <th>Location</th>
-              </tr>
-            </thead>
-            <tbody>
-              {jobs.map(job => {
-                return (
-                  <tr key={job.id}>
-                    <td>{job.date}</td>
-                    <td>{formatTime(job.startTime)}</td>
-                    <td>{job.client}</td>
-                    <td>{job.location}</td>
-                    <td>
-                      <button
-                        type="button"
-                        onClick={() => {
-                          handleRemove(job.id);
-                        }}
-                        className="btn btn-small btn-danger">
-                        Remove
-                      </button>
-                    </td>
-                  </tr>
-                );
-              })}
-            </tbody>
-          </Table>
-        </div>
-      </div>
-    </div>
-  );
-};
+import { firestore } from "../../../firebase";
+import { collectIdsAndDocs } from "../../../utilities";
+import FutureView from "./FutureView";
 
 export default class Future extends React.Component {
   constructor(props) {
@@ -62,8 +15,10 @@ export default class Future extends React.Component {
   unsubscribe = null;
 
   componentDidMount = async () => {
+    //temp uid - will get user info from global store when redux implemented
+    const uid = "user";
     this.unsubscribe = await firestore
-      .collection(`users/${this.props.location.state.uid}/jobs`)
+      .collection(`users/${uid}/jobs`)
       .onSnapshot(snapshot => {
         const jobs = snapshot.docs.map(collectIdsAndDocs);
         this.setState({ jobs });
@@ -75,9 +30,8 @@ export default class Future extends React.Component {
   };
 
   handleRemove = async id => {
-    await firestore
-      .doc(`users/${this.props.location.state.uid}/jobs/${id}`)
-      .delete();
+    await firestore;
+    const uid = "user".doc(`users/${uid}/jobs/${id}`).delete();
   };
 
   static propTypes = { fullView: PropTypes.bool.isRequired };
@@ -91,9 +45,9 @@ export default class Future extends React.Component {
       );
     } else {
       return (
-        <Col>
-          <FullTable jobs={this.state.jobs} handleRemove={this.handleRemove} />
-        </Col>
+        <div className="col">
+          <FutureView jobs={this.state.jobs} handleRemove={this.handleRemove} />
+        </div>
       );
     }
   }
